update show and _action,html.erb

+++ app/helpers/user_helper.rb
@@ -13,9 +13,8 @@ module UserHelper
   end
 
   def cancle(user)
-    unless current_user.friend_request_pending_to?(user)
-      link_to('cancle', user_friendships_path(user), method: :delete)
-    end
+    return link_to('cancle', user_friendships_path(user), method: :delete) if current_user.friend_request_pending_to?(user)
+    
   end
 
   def inivite(user)
